+++ lib/sendmsg.c
@@ -6,13 +6,13 @@
 
 extern int socketcall(int callno,long* args);
 
-int __libc_sendmsg(int a, const struct msghdr* msg, int flags);
-int __libc_sendmsg(int a, const struct msghdr* msg, int flags) {
+ssize_t __libc_sendmsg(int a, const struct msghdr* msg, int flags);
+ssize_t __libc_sendmsg(int a, const struct msghdr* msg, int flags) {
   long args[] = { a, (long) msg, flags };
   return socketcall(SYS_SENDMSG, args);
 }
 
-int sendmsg(int a, const struct msghdr *msg, int flags)
+ssize_t sendmsg(int a, const struct msghdr *msg, int flags)
  __attribute__ ((weak,alias("__libc_sendmsg"))) ;
 
 #endif
